+++ account/urls.py
@@ -1,0 +1,180 @@
+from django.urls import path,include
+from account.views import *
+
+
+urlpatterns = [
+
+
+    path('register/',UserCreateApiView.as_view(),name='register'),
+    path('login/',UserLoginAPIView.as_view(),name='login'),
+    path('password-reset/<uidb64>/<token>/',PasswordTokenCheckAPIView.as_view(),name='password-reset'),
+    path('pasword-reset-request/',PasswrodResetAPIView.as_view(),name='request-email'),
+    path('set-password/',SetNewPasswordAPIView.as_view(),name='set-new-passwrod'),
+
+]
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
