+++ tweet/admin.py
@@ -1,17 +1,21 @@
 from django.contrib import admin
-from .models import Tweet,TweetLike,Profile,Comment
+from tweet.models import Tweet, TweetLike, Comment
+from tweet.models import Profile
 
-# Register your models here.
+
 class TweetLikeAdmin(admin.TabularInline):
     model = TweetLike
+
 
 class TweetAdmin(admin.ModelAdmin):
     inlines = [TweetLikeAdmin]
     list_display = ['__str__','user']
     search_fields = ['content','user__username']
+
     class Meta:
         model =Tweet
 
+
 admin.site.register(Tweet,TweetAdmin)
-admin.site.register(Profile)
 admin.site.register(Comment)
+admin.site.register(Profile)
