+++ tweeter/settings.py
@@ -42,6 +42,7 @@ INSTALLED_APPS = [
     'rest_framework',
 #     internal apps
     'tweet',
+    'account'
 ]
 
 MIDDLEWARE = [
